Modularized Gradle plugins
Moved the complexity to separate plugins so that our code has better function decomposition (is easier to reuse, more flexible).
