+++ src/main/groovy/org/shipkit/internal/gradle/ShipkitJavaPlugin.java
@@ -1,20 +1,20 @@
 package org.shipkit.internal.gradle;
 
 import com.jfrog.bintray.gradle.BintrayExtension;
-import org.gradle.api.*;
+import org.gradle.api.Action;
+import org.gradle.api.Plugin;
+import org.gradle.api.Project;
+import org.gradle.api.Task;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
-import org.shipkit.gradle.UpdateReleaseNotesTask;
 import org.shipkit.gradle.ReleaseConfiguration;
+import org.shipkit.gradle.UpdateReleaseNotesTask;
+import org.shipkit.internal.gradle.release.ReleasePlugin;
 import org.shipkit.internal.gradle.util.BintrayUtil;
-import org.shipkit.internal.gradle.util.TaskMaker;
 
 import static org.shipkit.internal.gradle.BaseJavaLibraryPlugin.MAVEN_LOCAL_TASK;
-import static org.shipkit.internal.gradle.BaseJavaLibraryPlugin.POM_TASK;
-import static org.shipkit.internal.gradle.ContributorsPlugin.FETCH_ALL_CONTRIBUTORS_TASK;
 import static org.shipkit.internal.gradle.configuration.DeferredConfiguration.deferredConfiguration;
-import static org.shipkit.internal.gradle.configuration.LazyConfiguration.lazyConfiguration;
-import static org.shipkit.internal.gradle.util.Specs.withName;
+
 /**
  * Opinionated continuous delivery plugin.
  * Applies following plugins and preconfigures tasks provided by those plugins:
@@ -43,12 +43,10 @@ public class ShipkitJavaPlugin implements Plugin<Project> {
         //TODO ShipkitJavaPlugin should have no code but only apply other plugins
         //This way it will be easy for others to put together setup for other tools / build systems
 
-        project.getPlugins().apply(ReleaseNotesPlugin.class);
-        project.getPlugins().apply(AutoVersioningPlugin.class);
         project.getPlugins().apply(GitPlugin.class);
-        project.getPlugins().apply(ContributorsPlugin.class);
         project.getPlugins().apply(TravisPlugin.class);
-        project.getPlugins().apply(ReleaseNeededPlugin.class);
+        project.getPlugins().apply(PomContributorsPlugin.class);
+        project.getPlugins().apply(ReleasePlugin.class);
 
         project.allprojects(new Action<Project>() {
             @Override
@@ -62,38 +60,10 @@ public class ShipkitJavaPlugin implements Plugin<Project> {
                     });
                 }
 
-                subproject.getPlugins().withType(BaseJavaLibraryPlugin.class, new Action<BaseJavaLibraryPlugin>() {
-                    @Override
-                    public void execute(BaseJavaLibraryPlugin p) {
-                        final Task fetcher = project.getTasks().getByName(FETCH_ALL_CONTRIBUTORS_TASK);
-                        //Because maven-publish plugin uses new configuration model, we cannot get the task directly
-                        //So we use 'matching' technique
-                        subproject.getTasks().matching(withName(POM_TASK)).all(new Action<Task>() {
-                            public void execute(Task t) {
-                                t.dependsOn(fetcher);
-                            }
-                        });
-                    }
-                });
-            }
-        });
-
-        final Task performRelease = TaskMaker.task(project, "performRelease", new Action<Task>() {
-            public void execute(final Task t) {
-                t.setDescription("Performs release. " +
-                        "Ship with: './gradlew performRelease -Pshipkit.dryRun=false'. " +
-                        "Test with: './gradlew testRelease'");
-
-                t.dependsOn(VersioningPlugin.BUMP_VERSION_FILE_TASK, "updateReleaseNotes");
-                t.dependsOn(GitPlugin.PERFORM_GIT_PUSH_TASK);
-            }
-        });
-
-        project.allprojects(new Action<Project>() {
-            public void execute(final Project subproject) {
                 subproject.getPlugins().withType(JavaLibraryPlugin.class, new Action<JavaLibraryPlugin>() {
                     public void execute(JavaLibraryPlugin plugin) {
                         Task bintrayUpload = subproject.getTasks().getByName(BintrayPlugin.BINTRAY_UPLOAD_TASK);
+                        Task performRelease = project.getTasks().getByName(ReleasePlugin.PERFORM_RELEASE_TASK);
                         performRelease.dependsOn(bintrayUpload);
 
                         //Making git push run as late as possible because it is an operation that is hard to reverse.
@@ -118,38 +88,6 @@ public class ShipkitJavaPlugin implements Plugin<Project> {
                         });
                     }
                 });
-            }
-        });
-
-        TaskMaker.task(project, "testRelease", new Action<Task>() {
-            @Override
-            public void execute(final Task t) {
-                t.setDescription("Tests the release procedure and cleans up. Safe to be invoked multiple times.");
-                //releaseCleanUp is already set up to run all his "subtasks" after performRelease is performed
-                //releaseNeeded is used here only to execute the code paths in the release needed task (extra testing)
-                t.dependsOn("releaseNeeded", "performRelease", "releaseCleanUp");
-
-                //Ensure that when 'testRelease' is invoked we must be using 'dryRun'
-                //This is to avoid unintentional releases during testing
-                lazyConfiguration(t, new Runnable() {
-                    public void run() {
-                        if (!conf.isDryRun()) {
-                            throw new GradleException("When '" + t.getName() + "' task is executed" +
-                                    " 'shipkit.dryRun' must be set to 'true'.\n" +
-                                    "See Javadoc for ReleaseConfigurationPlugin.");
-                        }
-                    }
-                });
-            }
-        });
-
-        TaskMaker.task(project, "releaseCleanUp", new Action<Task>() {
-            public void execute(final Task t) {
-                t.setDescription("Cleans up the working copy, useful after dry running the release");
-
-                //using finalizedBy so that all clean up tasks run, even if one of them fails
-                t.finalizedBy(GitPlugin.PERFORM_GIT_COMMIT_CLEANUP_TASK);
-                t.finalizedBy(GitPlugin.TAG_CLEANUP_TASK);
             }
         });
     }
