+++ src/test/groovy/org/shipkit/internal/gradle/PomContributorsPluginTest.groovy
@@ -1,0 +1,15 @@
+package org.shipkit.internal.gradle
+
+import org.gradle.testfixtures.ProjectBuilder
+import spock.lang.Specification
+
+class PomContributorsPluginTest extends Specification {
+
+    def project = new ProjectBuilder().build()
+
+    def "applies"() {
+        expect:
+        project.plugins.apply(PomContributorsPlugin)
+        project.plugins.apply(BaseJavaLibraryPlugin)
+    }
+}
