+++ src/test/groovy/org/shipkit/internal/gradle/release/ReleasePluginTest.groovy
@@ -1,0 +1,14 @@
+package org.shipkit.internal.gradle.release
+
+import org.gradle.testfixtures.ProjectBuilder
+import spock.lang.Specification
+
+class ReleasePluginTest extends Specification {
+
+    def project = new ProjectBuilder().build()
+
+    def "applies"() {
+        expect:
+        project.plugins.apply(ReleasePlugin.class)
+    }
+}
